# baseline: Use custom instruction table