Previously EVMC instruction tables were used (for quick head start).
Now two EVMC tables are replaced with single one. This at least cuts memory bandwidth from 12 bytes to 4 bytes per instruction.
Changes to instruction requirements checks are minimal but the stack overflow is checked differently.
The check can be further optimized but this will be done separately.