@chfast 2021-06-07
Haswell 4.4 GHz, clang-12
```
Comparing o/b-master-v2 to o/b-itable-0
Benchmark                                                                Time             CPU      Time Old      Time New       CPU Old       CPU New
-----------------------------------------------------------------------------------------------------------------------------------------------------
baseline/analyse/main/blake2b_huff_mean                               -0.0764         -0.0765             5             5             5             5
baseline/execute/main/blake2b_huff/empty_mean                         -0.0835         -0.0835            22            20            22            20
baseline/execute/main/blake2b_huff/2805nulls_mean                     -0.0953         -0.0953           347           314           347           314
baseline/execute/main/blake2b_huff/5610nulls_mean                     -0.1003         -0.1003           674           607           674           607
baseline/execute/main/blake2b_huff/8415nulls_mean                     -0.0933         -0.0933           980           889           980           889
baseline/execute/main/blake2b_huff/65536nulls_mean                    -0.1004         -0.1004          7638          6871          7638          6871
baseline/analyse/main/blake2b_shifts_mean                             -0.0119         -0.0119             3             3             3             3
baseline/execute/main/blake2b_shifts/2805nulls_mean                   -0.0566         -0.0566          3422          3229          3422          3229
baseline/execute/main/blake2b_shifts/5610nulls_mean                   -0.0536         -0.0536          7011          6635          7011          6635
baseline/execute/main/blake2b_shifts/8415nulls_mean                   -0.0529         -0.0529         10515          9958         10515          9958
baseline/execute/main/blake2b_shifts/65536nulls_mean                  -0.0521         -0.0521         83230         78892         83229         78891
baseline/analyse/main/sha1_divs_mean                                  +0.0124         +0.0124             1             1             1             1
baseline/execute/main/sha1_divs/empty_mean                            -0.0109         -0.0109            56            55            56            55
baseline/execute/main/sha1_divs/1351_mean                             -0.0234         -0.0234          1137          1110          1137          1111
baseline/execute/main/sha1_divs/2737_mean                             -0.0253         -0.0253          2219          2163          2219          2163
baseline/execute/main/sha1_divs/5311_mean                             -0.0248         -0.0248          4330          4223          4330          4223
baseline/execute/main/sha1_divs/65536_mean                            -0.0261         -0.0261         52758         51382         52759         51383
baseline/analyse/main/sha1_shifts_mean                                -0.0073         -0.0073             1             1             1             1
baseline/execute/main/sha1_shifts/empty_mean                          -0.0563         -0.0563            35            33            35            33
baseline/execute/main/sha1_shifts/1351_mean                           -0.0573         -0.0573           710           670           710           670
baseline/execute/main/sha1_shifts/2737_mean                           -0.0487         -0.0487          1389          1321          1389          1321
baseline/execute/main/sha1_shifts/5311_mean                           -0.0582         -0.0582          2701          2544          2701          2544
baseline/execute/main/sha1_shifts/65536_mean                          -0.0577         -0.0577         32940         31041         32940         31041
baseline/analyse/main/weierstrudel_mean                               -0.0491         -0.0491             7             7             7             7
baseline/execute/main/weierstrudel/0_mean                             +0.0070         +0.0070           179           180           179           180
baseline/execute/main/weierstrudel/1_mean                             +0.0025         +0.0024           393           394           393           394
baseline/execute/main/weierstrudel/3_mean                             +0.0013         +0.0013           610           611           610           611
baseline/execute/main/weierstrudel/9_mean                             +0.0009         +0.0009          1260          1262          1260          1262
baseline/execute/main/weierstrudel/14_mean                            -0.0014         -0.0014          1805          1802          1805          1802
```