@chfast 2021-06-16
New benchmarks for execute only:
```
Comparing o/b-master to o/b-itable
Benchmark                                                                Time             CPU      Time Old      Time New       CPU Old       CPU New
-----------------------------------------------------------------------------------------------------------------------------------------------------
baseline/execute/main/blake2b_huff/empty_mean                         -0.0347         -0.0347            15            15            15            15
baseline/execute/main/blake2b_huff/2805nulls_mean                     -0.0385         -0.0386           330           317           330           317
baseline/execute/main/blake2b_huff/5610nulls_mean                     -0.0231         -0.0231           641           626           641           626
baseline/execute/main/blake2b_huff/8415nulls_mean                     -0.0304         -0.0304           939           910           939           910
baseline/execute/main/blake2b_huff/65536nulls_mean                    -0.0336         -0.0336          7300          7055          7300          7055
baseline/execute/main/blake2b_shifts/2805nulls_mean                   -0.0363         -0.0363          3160          3045          3160          3045
baseline/execute/main/blake2b_shifts/5610nulls_mean                   -0.0390         -0.0390          6278          6033          6278          6033
baseline/execute/main/blake2b_shifts/8415nulls_mean                   -0.0435         -0.0435          9426          9016          9426          9016
baseline/execute/main/blake2b_shifts/65536nulls_mean                  -0.0432         -0.0432         72934         69784         72934         69784
baseline/execute/main/sha1_divs/empty_mean                            -0.0150         -0.0150            54            53            54            53
baseline/execute/main/sha1_divs/1351_mean                             -0.0137         -0.0137          1124          1109          1124          1109
baseline/execute/main/sha1_divs/2737_mean                             -0.0126         -0.0126          2196          2168          2196          2168
baseline/execute/main/sha1_divs/5311_mean                             -0.0096         -0.0096          4273          4232          4273          4232
baseline/execute/main/sha1_divs/65536_mean                            -0.0132         -0.0132         52287         51598         52287         51598
baseline/execute/main/sha1_shifts/empty_mean                          -0.0222         -0.0222            33            32            33            32
baseline/execute/main/sha1_shifts/1351_mean                           -0.0184         -0.0184           696           684           696           684
baseline/execute/main/sha1_shifts/2737_mean                           -0.0180         -0.0180          1357          1333          1357          1333
baseline/execute/main/sha1_shifts/5311_mean                           -0.0213         -0.0213          2655          2598          2655          2598
baseline/execute/main/sha1_shifts/65536_mean                          -0.0203         -0.0203         32332         31677         32332         31677
baseline/execute/main/weierstrudel/0_mean                             -0.0389         -0.0389           169           163           169           163
baseline/execute/main/weierstrudel/1_mean                             -0.0215         -0.0215           380           371           380           371
baseline/execute/main/weierstrudel/3_mean                             -0.0214         -0.0214           594           582           594           582
baseline/execute/main/weierstrudel/9_mean                             -0.0210         -0.0210          1235          1209          1235          1209
baseline/execute/main/weierstrudel/14_mean                            -0.0216         -0.0216          1771          1733          1771          1733
baseline/execute/micro/beginsub_push1s_0xffff_mean                    -0.0009         -0.0009             0             0             0             0
baseline/execute/micro/beginsubs_0xffff_mean                          +0.0006         +0.0006             0             0             0             0
baseline/execute/micro/jumpdests_0xffff_mean                          -0.1046         -0.1046           108            97           108            97
baseline/execute/micro/loop_with_many_jumpdests_mean                  -0.1177         -0.1177         13273         11711         13273         11711
baseline/execute/micro/push1s_0xffff_mean                             -0.0001         -0.0001             0             0             0             0
baseline/execute/micro/push32s_0xffff_mean                            -0.0000         -0.0000             0             0             0             0
baseline/execute/micro/zeros_0xffff_mean                              +0.0005         +0.0005             0             0             0             0
```